Added Vulkan Instance and creating a logical device

+++ src/Core/Timestep.h
@@ -1,17 +1,20 @@
 #pragma once
+
+#include "Definitions.h"
+
 class Timestep {
 public:
     Timestep() = default;
-    Timestep(double time) : mTime(time) {}
-    double toMillis() const { return mTime * (double)1000; }
-    double toMicros() const { return mTime * (double)1000000; }
-    double Time() const { return mTime; }
+    Timestep(f32 time) : mTime(time) {}
+    f32 toMillis() const { return mTime * (f32)1000; }
+    f32 toMicros() const { return mTime * (f32)1000000; }
+    f32 Time() const { return mTime; }
     Timestep operator-(Timestep& other) {
         return Timestep(this->Time() - other.Time());
     }
-    void setTime(double time) {
+    void setTime(f32 time) {
         mTime = time;
     }
 private:
-    double mTime{};
+    f32 mTime{};
 };
